WIP: Remove "Le Reset"
The hackerspace announced in June it would not reopen in September: https://twitter.com/le_RESET/status/1271428089319432193 

However, the thread say they are going to announce more in a few days, hence the WIP tag. I am not sure how people want to handle deprecation on the index page, but a PR seemed fitting as a way to discuss that.